Video Recording: Big revamp to improve overall performance. We now don't rely on the periodic display refresh callback anymore to render the entire framebuffer but now rely on delta updates ("dirty rectangles"). Also, we now only encode new frames when an area has changed. This also needed cursor position + change change notifications, as we render the cursor on the host side if mouse integration is enabled (requires 7.1 Guest Additions as of now). Optimized the BGRA32->YUV IV420 color space conversion as well as the overall amount of pixel data shuffled forth and back. Added a new testcase for the cropping/centering code [build fixes]. bugref:10650
svn:sync-xref-src-repo-rev: r163627

+++ src/VBox/Main/src-client/RecordingUtils.cpp
@@ -1,4 +1,4 @@
-/* $Id: RecordingUtils.cpp 105006 2024-06-24 17:43:00Z andreas.loeffler@oracle.com $ */
+/* $Id: RecordingUtils.cpp 105009 2024-06-24 17:57:54Z andreas.loeffler@oracle.com $ */
 /** @file
  * Recording utility code.
  */
@@ -435,7 +435,7 @@ int RecordingUtilsDbgDumpImageData(const uint8_t *pu8RGBBuf, size_t cbRGBBuf, co
         RTFileWrite(fh, &coreHdr, sizeof(coreHdr), NULL);
 
         /* Bitmaps (DIBs) are stored upside-down (thanks, OS/2), so work from the bottom up. */
-        uint32_t offSrc = cbRGBBuf - uBytesPerLine;
+        size_t offSrc = cbRGBBuf - uBytesPerLine;
 
         /* Do the copy. */
         for (unsigned int i = 0; i < uHeight; i++)
